.claude/worktrees/jolly-wiles/requirements.txt
# Jarvis Assistant - Dependencies
# Install: pip install -r requirements.txt

# System tray (no console)
pystray>=0.19.0
Pillow>=9.0.0

# Speech
SpeechRecognition>=3.10.0
PyAudio>=0.2.13
pyttsx3>=2.90

# AI APIs
openai>=1.0.0
google-generativeai>=0.3.0

# Desktop automation: DOM first (Playwright), UI Automation (pywinauto), keyboard fallback
pygetwindow>=0.0.9
pyautogui>=0.9.54
playwright>=1.40.0
pywinauto>=0.6.8

# Global hotkey (F8 to exit)
keyboard>=0.13.5

# Windows APIs (needed by pystray/keyboard for EXE build; avoids "ordinal was not found")
pywin32>=306; sys_platform == "win32"
